Fixing style check errors

+++ src/common_util.h
@@ -32,8 +32,8 @@
 #include "parser.h"
 
 #ifdef WIN32
-#include <regex>
-typedef long long suseconds_t;
+#include <regex>  // NOLINT [build/c++11]
+typedef int64_t suseconds_t;
 inline std::string FormatPath(const std::string &filename) {
   // If the filename path comes from a bash command, translate its posix drive
   //  letter into a windows one
@@ -103,8 +103,11 @@ class Timer {
 
     suseconds_t diff_microsec = timeval2microsec(end_tv_) -
                                 timeval2microsec(start_tv_);
-    struct timeval diff_tv = {.tv_sec = static_cast<decltype(diff_tv.tv_sec)>(diff_microsec / kMicroSecs),
-                              .tv_usec = static_cast<decltype(diff_tv.tv_usec)>(diff_microsec % kMicroSecs)};
+    struct timeval diff_tv = {
+      .tv_sec  =
+        static_cast<decltype(diff_tv.tv_sec)>(diff_microsec / kMicroSecs),
+      .tv_usec =
+        static_cast<decltype(diff_tv.tv_usec)>(diff_microsec % kMicroSecs)};
     return diff_tv;
   }
